refactor: move files into models and commands packages
