+++ src/main/java/commands/Command.java
@@ -1,3 +1,5 @@
+package commands;
+
 public interface Command {
     void execute();
 }
