+++ src/main/java/commands/UnmarkCommand.java
@@ -1,16 +1,17 @@
+package commands;
+
+import commands.parser.Parser;
+import models.Task;
+
 import java.util.List;
 
-public class MarkCommand implements Command {
+public class UnmarkCommand implements Command {
     private final int itemIndex;
     private final List<Task> items;
 
-    public MarkCommand(List<Task> items, String input) {
+    public UnmarkCommand(List<Task> items, String input) {
         String inputArgs = Parser.parseInput(input).args();
         int index = Integer.parseInt(inputArgs) - 1;
-
-        if (index < 0 || index >= items.size()) {
-            throw new InvalidIndexException(items.size(), index);
-        }
 
         this.items = items;
         this.itemIndex = index;
@@ -19,7 +20,7 @@ public class MarkCommand implements Command {
     @Override
     public void execute() {
         Task task = items.get(itemIndex);
-        task.markDone();
-        System.out.printf("Marked this task as done:\n%s\n", task);
+        task.unmarkDone();
+        System.out.printf("Marked this task as not yet done:\n%s\n", task);
     }
 }
