+++ src/main/java/models/Deadline.java
@@ -1,3 +1,5 @@
+package models;
+
 public class Deadline extends Task {
     private final String by;
     public Deadline(String name, String by) {
