+++ src/main/java/models/Event.java
@@ -1,3 +1,5 @@
+package models;
+
 public class Event extends Task {
     private final String from;
     private final String to;
