+++ src/main/java/models/ToDo.java
@@ -1,3 +1,5 @@
+package models;
+
 public class ToDo extends Task {
     public ToDo(String name) {
         super(name);
